Debug prints changed
Changed a couple of the debugging print statements to be selected by
TM_STATUS rather than TM_STATUS_VERBOSE, so the former would be more
useful. This change was actually made in my code a couple of years ago
and never syncd. There are only about four lines changed, but I think
I’m using the mac client wrong, as it looks like it wants to replace
the entire .cpp and .h files.